Make DisableSpriteRenderingComponent attribute assembly targeted
Breaking changes!

* As InstancedProperty (InstancedPropertyComponent from now) DisableSpriteRenderingComponent (DisableRenderingComponent from now) is now assembly level, so client should declare it with [assembly: ...] out of namespaces and sign a type with attribute. The reason is also like with InstancedPropertyComponent  to let client registrate disable components with external assemblies types
* File/Types/Utils name refactoring
