--- Rendering/Common/DisableSpriteRenderingComponent.cs
@@ -1,7 +1,0 @@
-﻿using System;
-
-namespace NSprites
-{
-    [AttributeUsage(AttributeTargets.Struct, AllowMultiple = false)]
-    public class DisableSpriteRenderingComponent : Attribute { }
-}
